Fit an Isolation Forest

'solitude' class implements the isolation forest method introduced by paper Isolation based Anomaly Detection (Liu, Ting and Zhou <doi:10.1145/2133360.2133363>). The extremely randomized trees (extratrees) required to build the isolation forest is grown using ranger function from ranger package.

Design

$new() initiates a new 'solitude' object. The possible arguments are:

sample_size: (positive integer, default = 256) Number of observations in the dataset to used to build a tree in the forest

num_trees: (positive integer, default = 100) Number of trees to be built in the forest

replace: (boolean, default = FALSE) Whether the sample of observations should be chosen with replacement when sample_size is less than the number of observations in the dataset

seed: (positive integer, default = 101) Random seed for the forest

nproc: (NULL or a positive integer, default: NULL, means use all resources) Number of parallel threads to be used by ranger

respect_unordered_factors: (string, default: "partition")See respect.unordered.factors argument in ranger

$fit() fits a isolation forest for the given dataframe, computes depths of terminal nodes of each tree and stores the anomaly scores and average depth values in $scores object as a data.table

$predict() returns anomaly scores for a new data as a data.table

Details

Parallelization: ranger is parallelized and by default uses all the resources. This is supported when nproc is set to NULL. The process of obtaining depths of terminal nodes (which is excuted with $fit() is called) may be parallelized separately by setting up a future backend.

Methods

Public methods

isolationForest$new()

isolationForest$fit()

isolationForest$predict()

isolationForest$clone()

Method new()

Usage

isolationForest$new(
  sample_size = 256,
  num_trees = 100,
  replace = FALSE,
  seed = 101,
  nproc = NULL,
  respect_unordered_factors = NULL
)

Method fit()

Usage

isolationForest$fit(dataset)

Method predict()

Usage

isolationForest$predict(data)

Method clone()

The objects of this class are cloneable with this method.

Usage

isolationForest$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

data("humus", package = "mvoutlier")
columns_required = setdiff(colnames(humus)
                           , c("Cond", "ID", "XCOO", "YCOO", "LOI")
                           )
humus2 = humus[ , columns_required]
str(humus2)

#> 'data.frame':	617 obs. of  39 variables:
#>  $ Ag: num  0.124 0.13 0.124 0.309 0.04 0.421 0.235 0.043 0.07 0.086 ...
#>  $ Al: num  1420 7190 1970 6360 630 2060 1200 1510 3380 1320 ...
#>  $ As: num  0.743 0.973 1.1 1.24 0.533 1.54 0.836 0.657 1.75 1.04 ...
#>  $ B : num  2.51 3.45 1.62 1.59 4.33 1.93 2.51 4.46 5.08 2.1 ...
#>  $ Ba: num  74 65.2 121 63.5 36 111 109 24.5 42.2 51.4 ...
#>  $ Be: num  0.02 0.08 0.04 0.07 0.01 0.07 0.01 0.05 0.06 0.01 ...
#>  $ Bi: num  0.108 0.084 0.137 0.135 0.076 0.176 0.145 0.075 0.259 0.112 ...
#>  $ Ca: num  3900 1900 3240 2470 3500 3340 3010 3090 1760 2810 ...
#>  $ Cd: num  0.298 0.39 0.295 0.222 0.214 0.327 0.294 0.18 0.295 0.284 ...
#>  $ Co: num  1.28 3.37 0.98 3.91 0.43 4.61 0.75 0.55 1.68 1.12 ...
#>  $ Cr: num  3.04 2.02 1.33 17 0.79 3.78 2.01 1.63 10.9 1.71 ...
#>  $ Cu: num  10.9 12.3 5.59 29.8 6.23 47.3 8.66 5.55 10 9.25 ...
#>  $ Fe: num  2050 3170 1100 6180 790 3370 1150 1800 4820 970 ...
#>  $ Hg: num  0.155 0.263 0.168 0.175 0.197 0.239 0.233 0.217 0.238 0.229 ...
#>  $ K : num  1100 900 1000 900 700 1200 1000 600 800 1000 ...
#>  $ La: num  2.3 6.4 1.7 5.3 1.2 7 1.5 2.9 6.3 1.2 ...
#>  $ Mg: num  820 670 790 1180 1770 800 660 1750 820 570 ...
#>  $ Mn: num  202 22.8 70.3 118 24.4 86.1 1210 16.7 39.2 115 ...
#>  $ Mo: num  0.662 0.286 0.166 0.28 0.248 0.473 0.185 0.372 0.277 0.15 ...
#>  $ Na: num  40 120 30 150 410 60 10 320 110 50 ...
#>  $ Ni: num  9.41 14.8 5.05 59.5 1.98 78 5.7 2.19 11.7 13.2 ...
#>  $ P : num  743 1030 922 717 795 942 1120 790 779 745 ...
#>  $ Pb: num  15.9 13.9 19 16.3 14.3 21.7 15.2 11.6 284 14.1 ...
#>  $ Rb: num  8.14 2.82 4.45 8.13 2.67 5.23 4.94 2.33 4.07 5.35 ...
#>  $ S : num  1300 1950 1750 965 1860 1410 1880 2210 1260 1250 ...
#>  $ Sb: num  0.12 0.161 0.25 0.017 0.169 0.217 0.189 0.126 0.809 0.202 ...
#>  $ Sc: num  0.6 1.2 0.4 1.4 0.2 0.7 0.4 0.6 1.1 0.3 ...
#>  $ Si: num  630 640 580 690 440 530 650 490 630 560 ...
#>  $ Sr: num  22.2 34.4 43.4 29.2 46.4 67.9 20.7 59.1 28.8 23.2 ...
#>  $ Th: num  0.413 0.281 0.246 0.816 0.25 0.432 0.371 0.535 1.13 0.109 ...
#>  $ Tl: num  0.081 0.068 0.077 0.099 0.06 0.084 0.088 0.042 0.084 0.097 ...
#>  $ U : num  0.24 0.12 0.042 0.163 0.127 0.112 0.051 0.265 0.221 0.031 ...
#>  $ V : num  6.79 3.89 2.86 16.1 1.63 7.4 3.55 2.57 13.9 2.76 ...
#>  $ Y : num  0.8 2.4 0.5 1.9 0.8 1.6 0.5 1.8 1.7 0.4 ...
#>  $ Zn: num  59.1 18.1 67.6 43.3 41.4 46 103 21.5 23.2 38.5 ...
#>  $ C : num  39.9 47.5 44.2 19.4 45.8 47.8 45.7 47.3 29.5 46.8 ...
#>  $ H : num  5.5 6.8 6.3 3 6.1 5.8 6.5 6.4 4.4 6.5 ...
#>  $ N : num  1.2 1.5 1.5 0.7 1.5 1.3 1.7 1.5 0.7 1 ...
#>  $ pH: num  3.9 4.1 3.8 4 4.1 3.7 4 4.2 4 3.6 ...

set.seed(1)
index = sample(ceiling(nrow(humus2) * 0.5))
# initiate an isolation forest
iso = isolationForest$new(sample_size = length(index))
iso$fit(dataset = humus2[index, ])

#> INFO  [23:27:29.020] Building Isolation Forest ...  
#> INFO  [23:27:29.228] done 
#> INFO  [23:27:29.237] Computing depth of terminal nodes ...  
#> INFO  [23:27:29.685] done 
#> INFO  [23:27:29.756] Completed growing isolation forest 

iso$predict(humus2[index, ]) # scores for train data

#>       id average_depth anomaly_score
#>   1:   1         14.87     0.3789179
#>   2:   2         20.91     0.2554793
#>   3:   3         14.00     0.4010543
#>   4:   4         15.81     0.3563716
#>   5:   5         19.25     0.2847111
#>  ---                                
#> 305: 305         19.83     0.2741358
#> 306: 306         18.79     0.2933878
#> 307: 307         16.72     0.3358236
#> 308: 308          8.52     0.5734834
#> 309: 309          6.87     0.6386842

iso$predict(humus2[-index, ]) # scores for new data (50% sample)

#>       id average_depth anomaly_score
#>   1:   1         14.09     0.3987056
#>   2:   2         12.35     0.4466512
#>   3:   3         14.99     0.3759621
#>   4:   4         15.32     0.3679518
#>   5:   5         16.97     0.3303890
#>  ---                                
#> 304: 304         10.48     0.5046256
#> 305: 305         13.38     0.4176146
#> 306: 306          9.06     0.5536251
#> 307: 307         16.64     0.3375815
#> 308: 308         19.01     0.2892056